[F] Move start from toc to sections
