+++ client/src/backend/components/authoring/TOCList/TOCEntry.js
@@ -5,15 +5,11 @@ import Utility from "global/components/utility";
 import { Link } from "react-router-dom";
 import lh from "helpers/linkHandler";
 import { useTranslation } from "react-i18next";
-import Tooltip from "global/components/atomic/Tooltip";
-import { textsAPI } from "api";
-import { useApiCallback } from "hooks";
 import * as Styled from "./styles";
 
 export default function TOCEntry({
   entry,
   depth,
-  isStart,
   innerRef,
   draggableProps,
   dragHandleProps,
@@ -32,15 +28,6 @@ export default function TOCEntry({
 
   const editUrl = lh.link("backendTextTOCEntryEdit", textId, entry.id);
 
-  const updateText = useApiCallback(textsAPI.update);
-
-  const onSetStart = async id => {
-    await updateText(textId, {
-      attributes: { startTextSectionId: id }
-    });
-    // TODO: add error handling
-  };
-
   return (
     <Styled.Item ref={innerRef} {...draggableProps}>
       <Styled.Inner
@@ -49,15 +36,6 @@ export default function TOCEntry({
         $isTarget={active === entry.id}
       >
         <Styled.ButtonGroup>
-          <Tooltip
-            content={t("backend.forms.text_toc.start_tooltip_content")}
-            xOffset="-100px"
-            yOffset="43px"
-          >
-            <Styled.Button onClick={() => onSetStart(entry.data?.sectionId)}>
-              <Utility.IconComposer size={24} icon="playOutline24" />
-            </Styled.Button>
-          </Tooltip>
           <Styled.Button
             onClick={() => onDelete(entry.id)}
             aria-label={t("actions.delete")}
@@ -74,7 +52,6 @@ export default function TOCEntry({
         <Styled.TitleWrapper>
           {!!depth && <Styled.ChildLink icon="tocLink16" />}
           <Styled.Title>{entry.data ? entry.data.title : ""}</Styled.Title>
-          {isStart && <Styled.Tag>Start</Styled.Tag>}
         </Styled.TitleWrapper>
         <Styled.BG $isDragging={isDragging} />
       </Styled.Inner>
